Update ebay_links.py
This code scrapes the eBay homepage, finds the URLs of all the Featured Collections, then goes into each of those links and finds the URLs for all the individual products within each collection. 

However, the scraper, as currently constructed, only scrapes some of the links (specifically, it only scrapes the first 12 out of 17 for each featured collection), even though every link is found within the same HTML tags and has the same attributes. Looking at the HTML code, the only difference I found is that the first 12 links and the final 5 links are separated by that chunk of XML code that I have included at the bottom, under the variable name potential_problem. Is this code the reason my scraper is not scraping the last five?